pyproject.toml
[project]
name = "casbin-fastapi-decorator"
version = "0.1.0"
description = "Casbin authorization decorator factory for FastAPI"
readme = "README.md"
authors = [
    { name = "Neko1313", email = "nikita.ribalchencko@yandex.ru" }
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "fastapi-decorators>=0.5.0",
    "casbin>=1.36.0",
]

[project.optional-dependencies]
jwt = [
    "casbin-fastapi-decorator-jwt"
]
db = [
    "casbin-fastapi-decorator-db"
]

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[tool.uv.workspace]
members = [
    "packages/*",
    "examples/core",
    "examples/core-jwt",
    "examples/core-db",
]

[tool.uv.sources]
casbin-fastapi-decorator-jwt = { workspace = true }
casbin-fastapi-decorator-db = { workspace = true }


[tool.pytest.ini_options]
asyncio_mode = "auto"
python_files = ["test_*.py", "*_test.py", "act.py", "error.py"]
markers = [
    "integration: integration tests (real I/O, HTTP)",
    "unit: unit tests (no real I/O)",
    "permission_guard: PermissionGuard component tests",
    "access_subject: AccessSubject component tests",
]

[tool.bandit]
exclude_dirs = ["tests"]

[dependency-groups]
lint = [
    "bandit>=1.9.3",
    "ruff>=0.15.0",
    "ty>=0.0.15",
]
test = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "httpx>=0.28.0",
]
